scripts/requirements.txt
requests==2.32.3
pydantic==2.10.4
python-dotenv==1.0.1
rich==13.9.4
